feat: adding initial Hela
[Documentation]

`example`: curated list of example code including function to calculate the radians of a given angle, mean of list number using Hela Module

`euler.Euler.totient`
calculate the totient function for a given number. the totient function, also known as Euler's totient function, is a function that counts the number of positive integers less than or equal to a given integer that are relatively prime to that integer. For each prime number in the `primes` list, the function iterates over all numbers that are multiples of the prime number. if a number is divisible by the prime number, the totient of the number is set the product of the totient of the prime number and the prime number, otherwise, the totient of the number is set to the product of the totient of the prime number and the prime number minus 1

`math`
curated list of math including `factorial()` to calculate the factorial of a number, `radians()` are convert an angle from degress to radians, `avg_mean()` calculate the mean of a list of numbers, `sine()` calculate the sine of an angle

Signed-off-by: slowy07 <slowy.arfy@gmail.com>