chore: adding distribution Beta, Uniform
[Documentation]

continous distribution used to model phenomena that have no natural
boundaries, such as the height of a preson or the wight of an object.
formula: f(x) = 1 / (b - a)
where:
  - f(x): probability density function
  - x: the random variable
  - a: the minimum variable
  - b: the maximum variable
for the cumulative distribution function a continous Uniform distribution is F(x) = x - a / (b - a)
where:
  - f(x): is the cumulative distribution function
  - x: the random variable
  - a: the minimum variable
  - b: the maximum variable

beta distribution is a continuous probability distribution that is used
to model the probability of a random variable taking on a value between
0 and 1.

formula: f(x) = \frac{(x^{\alpha - 1})(1 - x)^{\beta - 1}}{B(\alpha, \beta)}
where:
  - f(x): the probability density function
  - x: the random variable
  - a and b: the parameter of distribution
  - B(\alpha, \beta) : the beta function
the beta function is a special function that is defined as follow:
B(\alpha, \beta) = \int_0^1 x^{\alpha - 1} (1 - x)^{\beta - 1} dx
beta distribution is a very versatile distribution that can be used to model a wide variety of phenomena. It is often used to model the uncertainty about a parameter in Bayesian statistics

Signed-off-by: slowy07 <slowy.arfy@gmail.com>